\name{Presymtoms}
\alias{Presymtoms}
\title{Predict model for severe COVID-19 patients}
\usage{
  Presymtoms(args)
}
\arguments{
\item{exp}{a expression profile, test indexes in columns, samples in rows.}
}
\value{a dataframe, samples in rows, columns : output of the predictor(s)}

\description{
  predict whether COVID-19 patients without symptoms on admission are severe
}
\author{liangyuan}
\keyword{internal}
\note{This is a contribution}
